Assignment 6 completed
