--- CS2233 - Data Structures/Assignments/Assignments(Future)/Assignment 6/correct_output.txt
@@ -1,13 +1,0 @@
--1
-2
-4
-9
-9 1 3 2 5 
-1 0
-3 5
-4 7
-2 8
-5 9
-4 2 3 4 
-11 2 3 4 1 
--1
