--- CS2233 - Data Structures/Assignments/Assignments(Future)/Assignment 6/input.txt
@@ -1,15 +1,0 @@
-N 5
-E 1 2 10 3 5
-E 2 3 2 5 1
-E 3 2 3 5 9 4 2
-E 4 5 6 1 7
-E 5 4 4
-? 2 1
-? 2 3
-? 5 4
-? 3 5
-P 1 5
-D 1
-P 2 4
-P 2 1
-? 1 4
