+++ CS2233 - Data Structures/Assignments/Assignments(Future)/Assignment 7/Metadata.h
@@ -1,0 +1,9 @@
+namespace DataStructure {
+    class Metadata {
+    public:
+        int length;
+        Metadata(int length) {
+            this -> length = length;
+        }
+    };
+}
