+++ CS2233 - Data Structures/Assignments/Assignments(Future)/Assignment 7/input.txt
@@ -1,0 +1,15 @@
+N 15
+U 1 3
+U 2 4
+U 3 6
+U 4 6
+U 10 1
+U 3 5
+U 5 8
+U 4 13
+U 4 12
+U 9 12
+U 7 14
+U 15 14
+U 11 7
+U 7 1
